Kurusaki/requirements.txt
python-dotenv
git+https://github.com/Rapptz/discord.py
requests
names
nekos.py
dnspython
pymongo
youtube-dl
RandomWords
PyNaCl
lavalink
aiohttp==3.6.3
chardet==3.0.4
cchardet
aiodns